requirements.txt
# Core scientific stack
pandas>=2.2.0
numpy>=1.26.0

# Optimization engine
gurobipy>=11.0.0

# Interface / GUI
PySimpleGUI>=5.0.0

# Plotting & image export
plotly>=5.24.0
kaleido>=0.2.1

# File IO & YAML
openpyxl>=3.1.2
PyYAML>=6.0.1

# Optional (for smooth CSV/Excel handling)
xlrd>=2.0.1
